requirements.txt
# Python Dependencies for Ethos Chatbot with Gemini API# Requirements for Agentic Reasoning System

# Install with: pip install -r requirements.txt

# Core dependencies

flask==3.0.0pandas>=1.3.0

google-generativeai==0.3.2numpy>=1.21.0

flask-cors==4.0.0

# Google Gemini API Integration
google-generativeai>=0.3.0

# For symbolic computation
sympy>=1.9

# For logging and configuration
python-dateutil>=2.8.2

# For CSV processing and data validation
pathlib2>=2.3.7; python_version<"3.4"

# Additional utilities (optional)
matplotlib>=3.5.0  # For performance visualization
seaborn>=0.11.0    # For advanced plotting
scipy>=1.7.0       # For statistical analysis

# Development and testing dependencies (optional)
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
